+++ test/cloud_replication_package_test.dart
@@ -1,29 +1,12 @@
 import 'dart:convert';
-//import 'package:convert/convert.dart';
 import 'dart:io';
-//import 'dart:math';
-// ignore: library_prefixes
-//import 'package:cloud_replication_package/src/replication_exception.dart';
-// ignore: library_prefixes
-//import 'package:encrypt/encrypt.dart' as Enc;
 
-//import 'package:cloud_replication_package/cloud_replication_package.dart';
-//import 'package:cloud_replication_package/src/cloud_models/event.dart';
-//import 'package:cloud_replication_package/cloud_replication_package.dart';
-//import 'package:cloud_replication_package/src/cloud_models/event.dart';
-//import 'package:cloud_replication_package/src/cloud_models/short_user.dart';
-//import 'package:cloud_replication_package/src/cloud_models/threat_weights.dart';
 import 'package:cloud_replication_package/cloud_replication_package.dart';
 import 'package:cloud_replication_package/src/service/cloud_service/cloud_exception.dart';
-//import 'package:cloud_replication_package/src/cloud_models/user.dart';
-//import 'package:cloud_replication_package/src/service/cloud_service.dart';
-//import 'package:cloud_replication_package/src/service/event_listener.dart';
 import 'package:cloud_replication_package/src/service/node_listener.dart';
-import 'package:cloud_replication_package/src/service/replication_service.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:http/http.dart' as http;
 import 'package:http/io_client.dart';
-//import 'package:intl/intl.dart';
 import 'package:test/test.dart';
 import 'package:geiger_localstorage/geiger_localstorage.dart' as toolbox_api;
 import 'package:geiger_api/geiger_api.dart';
@@ -102,9 +85,6 @@ Future<GeigerApi> _initGeigerApi() async {
 Future<toolbox_api.StorageController> init() async {
   //print('[REPLICATION] INIT GEIGER STORAGE');
   WidgetsFlutterBinding.ensureInitialized();
-
-  //await toolbox_api.StorageMapper.initDatabaseExpander();
-  //storageController = toolbox_api.GenericController('Cloud-Replication', toolbox_api.SqliteMapper('dbFileName.bd'));
   return toolbox_api.GenericController(
       'Cloud-Replication', toolbox_api.DummyMapper('Cloud-Replication'));
 }
@@ -127,20 +107,26 @@ Future<String> generateUUID() async {
 }
 
 void replicationTests() async {
-  test("s", () async {
+  test("TEST UPDATE GLOBAL DATA - PROFILES & THREATS", () async {
     ReplicationController rep = ReplicationService();
     await rep.initGeigerStorage();
     await rep.updateThreatWeights();
-  });
-  test('STORAGE LISTENER - CHECK CODE', () async {
-    print("CHECK STORAGE LISTENER");
-    //Function? handler;
+    // GET :Global:threats
+    // GET :Global:profiles
     GeigerApi localMaster =
         (await getGeigerApi("", GeigerApi.masterId, Declaration.doShareData))!;
     // ignore: unused_local_variable
     toolbox_api.StorageController storageController = localMaster.getStorage()!;
-    //toolbox_api.SearchCriteria sc = toolbox_api.SearchCriteria(searchPath: ':');
-    //NodeListener stListener = NodeListener();
+    toolbox_api.Node profile = await storageController.get(":Global:profiles");
+    print("PROFILE NODE");
+    print(profile);
+    print("PROFILE CHILDREN");
+    print(await profile.getChildren());
+    toolbox_api.Node threats = await storageController.get(":Global:threats");
+    print("THREAT NODE");
+    print(threats);
+    print("THREAT CHILDREN");
+    print(await threats.getChildren());
   });
   test('STORAGE LISTENER - LAST NODE UPDATED', () async {
     print("LISTENER TEST - CHECKS DELETE BEHAVIOUR");
@@ -254,34 +240,6 @@ void replicationTests() async {
         print("CHECKER");
         await storageController.delete(entry.path);
       }
-    }
-  });*/
-  /*test('addOrUpdateTest', () async {
-    GeigerApi localMaster =
-        (await getGeigerApi("", GeigerApi.masterId, Declaration.doShareData))!;
-    toolbox_api.StorageController storageController = localMaster.getStorage()!;
-    try {
-      toolbox_api.Node first = await storageController.get(':Devices:newTestNode');
-      print(first);
-    } catch (e) {
-      print("CATCH - Node not found");
-      toolbox_api.Node nodeToAdd = toolbox_api.NodeImpl(':Devices:newTestNode', '');
-      await storageController.addOrUpdate(nodeToAdd);
-    }
-    try {
-      toolbox_api.Node checker = await storageController.get(':Devices:newTestNode');
-      print(checker);
-    } catch (e) {
-      print("NODE HAS NOT BEEN ADDED");
-    }
-    toolbox_api.Node nodeToAdd = toolbox_api.NodeImpl(':Devices:newTestNode', '');
-    nodeToAdd.addValue(toolbox_api.NodeValueImpl("demo", "demoTest"));
-    await storageController.addOrUpdate(nodeToAdd);
-    try {
-      toolbox_api.Node checker = await storageController.get(':Devices:newTestNode');
-      print(checker);
-    } catch (e) {
-      print("NODE HAS NOT BEEN ADDED");
     }
   });*/
   //late toolbox_api.StorageController storageController;