+++ CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.1 First Package
 
-* TODO: Describe initial release.
+Includes:
+- Full replication based on the defined strategies
+- Listeners to achieve event based replication
+- Pairing bugfixing
+- Replication of :Global data from Cloud to Local